requirements.txt
nextcord==2.6.0
aiosqlite==0.20.0
asyncio==3.4.3
python-dotenv==1.0.0
requests==2.31.0